requirements.txt
requests==2.26.0
pyTelegramBotAPI==3.7.9
python-dotenv==0.20.0